refactor: simplifications and cleanup
