+++ src/runtime/types.ts
@@ -1,7 +1,7 @@
 import type { NuxtApp } from '#app'
 import type { ComputedRef } from 'vue'
 import type { Directions, LocaleObject, Strategies } from '#internal-i18n-types'
-import type { Locale } from 'vue-i18n'
+import type { I18n, Locale } from 'vue-i18n'
 import type { RouteLocationNormalizedGeneric, RouteRecordNameGeneric } from 'vue-router'
 
 export type CompatRoute = Omit<RouteLocationNormalizedGeneric, 'name'> & {
@@ -168,6 +168,13 @@ export interface NuxtI18nRoutingCustomProperties<
    * Current locale properties.
    */
   localeProperties: LocaleObject
+}
+
+declare module '#app' {
+  interface NuxtApp {
+    /** @internal */
+    _vueI18n: I18n
+  }
 }
 
 declare module 'vue-i18n' {
